Prevent future import cycles by creating `collection` module.
This continues the reorganization by separating all the basic
`Collection`-related code into the `collection` module. This will
prevent an import cycle between `base` and `data` in a future commit.
