+++ python-spec/src/somacore/measurement.py
@@ -3,7 +3,7 @@
 from typing_extensions import Final
 
 from . import _wrap
-from . import base
+from . import collection
 from . import data
 
 
@@ -20,32 +20,32 @@ class Measurement(_wrap.CollectionProxy):
     All variables for this measurement _must_ be defined in this dataframe.
     """
 
-    X = _wrap.item(base.Collection[data.NDArray])
+    X = _wrap.item(collection.Collection[data.NDArray])
     """A collection of matrices containing feature values.
 
     Each matrix is indexed by ``[obsid, varid]``. Sparse and dense 2D arrays may
     both be used in any combination in ``X``.
     """
 
-    obsm = _wrap.item(base.Collection[data.DenseNDArray])
+    obsm = _wrap.item(collection.Collection[data.DenseNDArray])
     """Matrices containing annotations of each ``obs`` row.
 
     This has the same shape as ``obs`` and is indexed with ``obsid``.
     """
 
-    obsp = _wrap.item(base.Collection[data.SparseNDArray])
+    obsp = _wrap.item(collection.Collection[data.SparseNDArray])
     """Matrices containg pairwise annotations of each ``obs`` row.
 
     This is indexed by ``[obsid_1, obsid_2]``.
     """
 
-    varm = _wrap.item(base.Collection[data.DenseNDArray])
+    varm = _wrap.item(collection.Collection[data.DenseNDArray])
     """Matrices containing annotations of each ``var`` row.
 
     This has the same shape as ``var`` and is indexed with ``varid``.
     """
 
-    varp = _wrap.item(base.Collection[data.SparseNDArray])
+    varp = _wrap.item(collection.Collection[data.SparseNDArray])
     """Matrices containg pairwise annotations of each ``var`` row.
 
     This is indexed by ``[varid_1, varid_2]``.
